requirements.txt
psycopg2-binary
shioaji==1.1.3
python-dotenv==1.0.0